+++ src/f_midi.c
@@ -237,7 +237,7 @@ _WM_ParseNewMidi(const uint8_t *midi_data, uint32_t midi_size) {
         goto _end;
     }
 
-    if ((float)smallest_delta >= 0x7fffffff / samples_per_delta_f) {
+    if ((float)smallest_delta >= (float)0x7fffffff / samples_per_delta_f) {
         /* DEBUG */
         /* fprintf(stderr,"INTEGER OVERFLOW (samples_per_delta: %f, smallest_delta: %u)\n", */
         /*        samples_per_delta_f, smallest_delta); */
@@ -325,7 +325,7 @@ _WM_ParseNewMidi(const uint8_t *midi_data, uint32_t midi_size) {
             NEXT_TRACK: continue;
             }
 
-            if ((float)smallest_delta >= 0x7fffffff / samples_per_delta_f) {
+            if ((float)smallest_delta >= (float)0x7fffffff / samples_per_delta_f) {
                 /* DEBUG */
                 /* fprintf(stderr,"INTEGER OVERFLOW (samples_per_delta: %f, smallest_delta: %u)\n", */
                 /*        samples_per_delta_f, smallest_delta); */
@@ -399,7 +399,7 @@ _WM_ParseNewMidi(const uint8_t *midi_data, uint32_t midi_size) {
                 tracks[i]++;
                 track_size[i]--;
 
-                if ((float)smallest_delta >= 0x7fffffff / samples_per_delta_f) {
+                if ((float)smallest_delta >= (float)0x7fffffff / samples_per_delta_f) {
                     /* DEBUG */
                     /* fprintf(stderr,"INTEGER OVERFLOW (samples_per_delta: %f, smallest_delta: %u)\n", */
                     /*        samples_per_delta_f, smallest_delta); */
